+++ src/database/models.py
@@ -163,7 +163,6 @@ class UserModel(BaseModel):
     new_quest_coin_quantity: int = 2
     max_items_count_in_market: int = 4
     luck: int = 1
-    adverts_count: int = 0
     last_active_time: datetime = field(default_factory=_utcnow)
     achievement_progress: dict = field(default_factory=dict)
     accepted_rules: bool = False