+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Добавлено
 
 - Добавлена новая зависимость: `tinylogging`
+- В место стикера теперь полезные подсказки [#60](https://github.com/HamletSargsyan/livebot/issues/60)
 
 ### Изменено